+++ CHANGES.md
@@ -23,6 +23,7 @@ This files lists the changes during the lifetime of this project.
 - 162: Add GitHub action for labeling pre-commit PRs with label dependencies
 - 162: Change formatting according to styling rules
 - 161: Fix that BM link no longer resets active filters
+- 163: Generalize assignment import to be brand independent
 
 ## 2026-01-19